+++ app/javascript/vue/components/georeferences/geolocateComponent.vue
@@ -61,7 +61,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-  /deep/ .modal-container {
+  ::v-deep .modal-container {
     max-width: 500px;
   }
 </style>