+++ app/javascript/vue/routes/endpoints/ObservationMatrix.js
@@ -1,0 +1,11 @@
+import baseCRUD from './base'
+
+const permitParams = {
+  observation_matrix: {
+    name: String
+  }
+}
+
+export const ObservationMatrix = {
+  ...baseCRUD('observation_matrices', permitParams)
+}