+++ app/javascript/vue/routes/endpoints/Otu.js
@@ -1,0 +1,14 @@
+import baseCRUD, { annotations } from './base'
+
+const permitParams = {
+  otu: {
+    id: Number,
+    name: String,
+    taxon_name_id: Number
+  }
+}
+
+export const Otu = {
+  ...baseCRUD('otus', permitParams),
+  ...annotations('otus')
+}