+++ app/javascript/vue/routes/endpoints/Topic.js
@@ -1,0 +1,9 @@
+import baseCRUD from './base'
+
+const permitParams = {
+  topic: {}
+}
+
+export const Topic = {
+  ...baseCRUD('topics', permitParams)
+}