+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/components/verbatim/Latitude.vue
@@ -4,15 +4,27 @@
     <input
       class="full_width"
       type="text"
-      v-model="collectingEvent.verbatim_latitude">
+      v-model="latitude">
   </div>
 </template>
 
 <script>
 
 import extendCE from '../mixins/extendCE'
+import { parseCoordinateCharacters } from 'helpers/georeferences'
 
 export default {
-  mixins: [extendCE]
+  mixins: [extendCE],
+
+  computed: {
+    latitude: {
+      get () {
+        return this.collectingEvent.verbatim_latitude
+      },
+      set (value) {
+        this.collectingEvent.verbatim_latitude = parseCoordinateCharacters(value)
+      }
+    }
+  }
 }
 </script>