+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/store/actions/saveCollectingEvent.js
@@ -1,4 +1,4 @@
-import { CreateCollectingEvent, UpdateCollectingEvent } from '../../request/resources'
+import { CollectingEvent } from 'routes/endpoints'
 import { ActionNames } from '../actions/actions'
 import { MutationNames } from '../mutations/mutations'
 import { RouteNames } from 'routes/routes'
@@ -6,7 +6,6 @@ import SetParam from 'helpers/setParam'
 
 export default async ({ state, commit, dispatch }) => {
   const collectingEvent = state.collectingEvent
-  const saveCE = collectingEvent.id ? UpdateCollectingEvent : CreateCollectingEvent
 
   if (collectingEvent.units === 'ft') {
     ['minimum_elevation', 'maximum_elevation', 'elevation_precision'].forEach(key => {
@@ -18,7 +17,10 @@ export default async ({ state, commit, dispatch }) => {
 
   state.settings.isSaving = true
 
-  saveCE(collectingEvent).then(async response => {
+  return (collectingEvent.id
+    ? CollectingEvent.update(collectingEvent.id, { collecting_event: collectingEvent })
+    : CollectingEvent.create({ collecting_event: collectingEvent })
+  ).then(async response => {
     TW.workbench.alert.create(`Collecting event was successfully ${collectingEvent.id ? 'updated' : 'created'}.`, 'notice')
     commit(MutationNames.SetCollectingEvent, Object.assign({}, collectingEvent, response.body, { roles_attributes: response.body.collector_roles || [] }))
     await dispatch(ActionNames.ProcessGeoreferenceQueue)